feat: improve scroll behavior throughtout the app
