+++ lib/core/utils/custom_scroll_physics.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/widgets.dart';
+
+class CustomPageViewScrollPhysics extends ScrollPhysics {
+  const CustomPageViewScrollPhysics({super.parent});
+
+  @override
+  CustomPageViewScrollPhysics applyTo(ScrollPhysics? ancestor) {
+    return CustomPageViewScrollPhysics(parent: buildParent(ancestor));
+  }
+
+  @override
+  SpringDescription get spring => const SpringDescription(
+        mass: 200,
+        stiffness: 200,
+        damping: 0.8,
+      );
+}
